Adds the constant term check to subdivision solver
The constant term check compares the absolute value of the constant term
of each chebyshev polynomial to the sum of the absolute values of the
non-constant terms. If the absolute value of the constant term is
larger for each of them, this signifies that the zero locus for this
chebyshev polynomial in this interval is the empty set, and thus there
are no zeros on that interval, so the code will throw the interval out.

Testing has shown that the code runs faster when it includes a constant
term check than when it does not. Further testing will have to be done
to verify that the best time to run the constant term check is only
after each subdivision.