Pushed code for quadratic term check
Created a new file called QuadraticCheck.py. Works similar to the old
quadratic term check in the interval checks file, but slight changes
needed to be made in order for it to be compatible with new code. Now
it goes through each subinterval one by one and loops through each of
the functions, and if one of them is guaranteed not to have a root in
the interval according to the quadratic term check, it throws the
interval out.

Also added three new parameters to the solve function in
Combined_Solver.py: constant_check, low_dim_quadratic_check, and
all_dim_quadratic_check. The first runs the constant check after every
subdivision if set to true, the second runs the quadratic check after
every subdivision if set to true for dimensions 3 and lower, the third
runs the quadratic check after every subdivision for all dimensions if
set to true. The first two are true by default and the third is false by
default. This is because testing indicates that this is the fastest way
to do things.